Consolidating hv-route functionality
