+++ src/core/components/hv-route/index.tsx
@@ -6,7 +6,10 @@
  *
  */
 
+import * as Contexts from 'hyperview/src/contexts';
 import * as DomService from 'hyperview/src/services/dom';
+import * as Helpers from 'hyperview/src/services/dom/helpers-legacy';
+import * as Namespaces from 'hyperview/src/services/namespaces';
 import * as NavigationContext from 'hyperview/src/contexts/navigation';
 import * as NavigatorContext from 'hyperview/src/contexts/navigator';
 import * as NavigatorService from 'hyperview/src/services/navigator';
@@ -14,12 +17,30 @@ import * as Types from './types';
 import * as TypesLegacy from 'hyperview/src/types-legacy';
 import * as UrlService from 'hyperview/src/services/url';
 import React, { PureComponent, useContext } from 'react';
+import HvNavigator from 'hyperview/src/core/components/hv-navigator';
+import HvScreen from 'hyperview/src/core/components/hv-screen';
 import LoadError from 'hyperview/src/core/components/load-error';
 import Loading from 'hyperview/src/core/components/loading';
+import { RouteProps } from 'hyperview/src/core/components/hv-screen/types';
 
 type State = {
   doc: TypesLegacy.Document | null;
   error: Error | null;
+};
+
+type BuildScreenProps = {
+  doc: TypesLegacy.Document;
+  navLogic: NavigatorService.Navigator;
+  routeProps: Types.InnerRouteProps;
+  url: string | null;
+};
+
+type RouteRenderProps = {
+  doc: TypesLegacy.Document;
+  element?: TypesLegacy.Element;
+  navLogic: NavigatorService.Navigator;
+  routeProps: Types.InnerRouteProps;
+  url: string;
 };
 
 /**
@@ -92,7 +113,7 @@ class HvRouteInner extends PureComponent<Types.InnerRouteProps, State> {
   /**
    * View shown while loading
    */
-  LoadingView = (): React.ReactElement => {
+  LoadingComponent = (): React.ReactElement => {
     const LoadingScreen = this.props.loadingScreen || Loading;
     return <LoadingScreen />;
   };
@@ -100,7 +121,7 @@ class HvRouteInner extends PureComponent<Types.InnerRouteProps, State> {
   /**
    * View shown when there is an error
    */
-  ErrorView = (props: { error: unknown }): React.ReactElement => {
+  ErrorComponent = (props: { error: unknown }): React.ReactElement => {
     const ErrorScreen = this.props.errorScreen || LoadError;
     return (
       <ErrorScreen
@@ -115,9 +136,128 @@ class HvRouteInner extends PureComponent<Types.InnerRouteProps, State> {
   };
 
   /**
+   * Build the <HvScreen> component with injected props
+   */
+  ScreenComponent = (props: BuildScreenProps): React.ReactElement => {
+    // Inject the corrected url into the params and cast as correct type
+    const route: RouteProps = {
+      ...props.routeProps.route,
+      key: props.routeProps.route?.key || 'hv-screen',
+      name: props.routeProps.route?.name || 'hv-screen',
+      params: {
+        ...props.routeProps.route?.params,
+        url: props.url || undefined,
+      },
+    };
+
+    return (
+      <Contexts.DateFormatContext.Consumer>
+        {formatter => (
+          <HvScreen
+            back={props.navLogic.back}
+            behaviors={props.routeProps.behaviors}
+            closeModal={props.navLogic.closeModal}
+            components={props.routeProps.components}
+            doc={props.doc}
+            elementErrorComponent={props.routeProps.elementErrorComponent}
+            entrypointUrl={props.routeProps.entrypointUrl}
+            errorScreen={props.routeProps.errorScreen}
+            fetch={props.routeProps.fetch}
+            formatDate={formatter}
+            loadingScreen={props.routeProps.loadingScreen}
+            navigate={props.navLogic.navigate}
+            navigation={props.routeProps.navigation}
+            onParseAfter={props.routeProps.onParseAfter}
+            onParseBefore={props.routeProps.onParseBefore}
+            openModal={props.navLogic.openModal}
+            push={props.navLogic.push}
+            route={route}
+            url={props.url || undefined}
+          />
+        )}
+      </Contexts.DateFormatContext.Consumer>
+    );
+  };
+
+  /**
+   * Evaluate the <doc> element and render the appropriate component
+   */
+  RouteComponent = (props: RouteRenderProps): React.ReactElement => {
+    let renderElement: TypesLegacy.Element | null = null;
+
+    if (props.element) {
+      renderElement = props.element;
+    } else {
+      // Get the <doc> element
+      const root: TypesLegacy.Element | null = Helpers.getFirstTag(
+        props.doc,
+        TypesLegacy.LOCAL_NAME.DOC,
+      );
+      if (!root) {
+        throw new NavigatorService.HvRenderError('No root element found');
+      }
+
+      // Get the first child as <screen> or <navigator>
+      const screenElement: TypesLegacy.Element | null = Helpers.getFirstTag(
+        root,
+        TypesLegacy.LOCAL_NAME.SCREEN,
+      );
+      const navigatorElement: TypesLegacy.Element | null = Helpers.getFirstTag(
+        root,
+        TypesLegacy.LOCAL_NAME.NAVIGATOR,
+      );
+
+      if (!screenElement && !navigatorElement) {
+        throw new NavigatorService.HvRenderError(
+          'No <screen> or <navigator> element found',
+        );
+      }
+      renderElement = screenElement || navigatorElement;
+    }
+
+    if (!renderElement) {
+      throw new NavigatorService.HvRenderError('No element found');
+    }
+
+    if (renderElement.namespaceURI !== Namespaces.HYPERVIEW) {
+      throw new NavigatorService.HvRenderError('Invalid namespace');
+    }
+
+    if (renderElement.localName === TypesLegacy.LOCAL_NAME.NAVIGATOR) {
+      return <HvNavigator element={renderElement} />;
+    }
+    const { ScreenComponent } = this;
+
+    if (renderElement.localName === TypesLegacy.LOCAL_NAME.SCREEN) {
+      if (props.routeProps.handleBack) {
+        return (
+          <props.routeProps.handleBack>
+            <ScreenComponent
+              doc={props.doc}
+              navLogic={props.navLogic}
+              routeProps={props.routeProps}
+              url={props.url}
+            />
+          </props.routeProps.handleBack>
+        );
+      }
+      return (
+        <ScreenComponent
+          doc={props.doc}
+          navLogic={props.navLogic}
+          routeProps={props.routeProps}
+          url={props.url}
+        />
+      );
+    }
+
+    throw new NavigatorService.HvRenderError('Invalid element type');
+  };
+
+  /**
    * View shown when the document is loaded
    */
-  ContentView = (): React.ReactElement => {
+  ContentComponent = (): React.ReactElement => {
     if (!this.props.url) {
       throw new NavigatorService.HvRouteError('No url received');
     }
@@ -125,10 +265,10 @@ class HvRouteInner extends PureComponent<Types.InnerRouteProps, State> {
       throw new NavigatorService.HvRouteError('No document received');
     }
 
-    const { ErrorView } = this;
+    const { ErrorComponent, RouteComponent } = this;
     try {
       return (
-        <NavigatorService.Render
+        <RouteComponent
           doc={this.state.doc}
           element={this.props.element}
           navLogic={this.navLogic}
@@ -137,19 +277,19 @@ class HvRouteInner extends PureComponent<Types.InnerRouteProps, State> {
         />
       );
     } catch (err) {
-      return <ErrorView error={err} />;
+      return <ErrorComponent error={err} />;
     }
   };
 
   render() {
-    const { ErrorView, LoadingView, ContentView } = this;
+    const { ErrorComponent, LoadingComponent, ContentComponent } = this;
     if (this.state.error) {
-      return <ErrorView error={this.state.error} />;
+      return <ErrorComponent error={this.state.error} />;
     }
     if (this.state.doc) {
-      return <ContentView />;
+      return <ContentComponent />;
     }
-    return <LoadingView />;
+    return <LoadingComponent />;
   }
 }
